module: inherit unique lookup paths from parent
The previous commit added the module lookup paths based on a symlink
location as well as the realpath location.  However, _subsequent_
modules loaded in that environment would _not_ have those additional
paths added to their lookup path, since they are not loaded based on a
linked location.

This patch adds any paths from the parent module lookup paths to the
child, if the child does not already have that path in their lookup
list.